Refactor: feign으로 통신하는 로직 제거 -> 직접 호출로 변경
