--- gathering-service/src/main/java/com/sparta/moim/gathering/gathering/application/event/feign/InternalGatheringService.java
@@ -1,8 +1,0 @@
-package com.sparta.moim.gathering.gathering.application.event.feign;
-
-import java.util.UUID;
-
-public interface InternalGatheringService {
-  void validateGatheringExists(UUID gatheringId);
-  void validateGatheringStatusOpen(UUID gatheringId);
-}
