--- gathering-service/src/main/java/com/sparta/moim/gathering/gathering/application/event/feign/MemberService.java
@@ -1,10 +1,0 @@
-package com.sparta.moim.gathering.gathering.application.event.feign;
-
-import com.sparta.moim.gathering.gathering.application.dto.result.GetGatheringMemberListResult;
-import java.util.List;
-import java.util.UUID;
-
-public interface MemberService {
-  List<GetGatheringMemberListResult> findMembers(UUID gatheringId);
-
-}
