+++ gathering-service/src/main/java/com/sparta/moim/gathering/gathering/application/service/InternalMemberService.java
@@ -2,7 +2,7 @@ package com.sparta.moim.gathering.gathering.application.service;
 
 import com.sparta.moim.gathering.gathering.application.dto.result.GetMemberListResult;
 import com.sparta.moim.gathering.gathering.domain.entity.Member;
-import com.sparta.moim.gathering.gathering.domain.enums.MemberRepository;
+import com.sparta.moim.gathering.gathering.domain.repository.MemberRepository;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
